requirements.txt
dj-database-url==0.4.0
Django==1.9.2
django-material==0.7.0
djangorestframework==3.3.3
gunicorn==19.4.5
psycopg2==2.6.1
whitenoise==2.0.6
